Android: Add support for onRegion... events
onRegion events where already available on iOS, but it seemed not
on Android. It took me 2 days to figure it out why it was not
workning on it, as soon as I found out why, I tried to implement
it.

This commit adds support for:
* onRegionWillChange
* onRegionIsChanging
* onRegionDidChange

I've also added some TODO lines which I think it could be missing:
* Support for event off on all onRegion...
* Support for removing marker callback

Having that in mind, I've attached the callbacks trough instances
instead of class properties.

--
José Moreira